ESP32 compileability and Led library switch
This is very much WIP. With these changes, the code compiles on the
Adafruit QT Py ESP32-C3 - without BLE or storage functionalities.

This also switches the library that drives the LEDS to the Adafruit
NeoPixel library - to support RGBW pixels that are ill-supported by
FastLED. All color calculations are still done by fastled.